% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_matrix2lavaan}
\alias{check_matrix2lavaan}
\title{Checks matrices for matrix2lavaan function}
\arguments{
\item{m}{matrix, data.frame or tibble}

\item{mname}{Name of m}
}
\description{
Checks matrices for matrix2lavaan function
}
\keyword{internal}
