% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_model_names}
\alias{get_model_names}
\title{Return model names}
\usage{
get_model_names(m, ...)
}
\arguments{
\item{m}{Structural model represented by lavaan syntax or output of sim_standardized_matrices function.}

\item{...}{parameters passed to the `sim_standardized_matrices` function}
}
\value{
A list of variable names
}
\description{
Return model names
}
\examples{
m <- "
A =~ 0.5 * A1 + 0.8 * A2 + 0.8 * A3
B =~ 0.5 * B1 + 0.8 * B2 + 0.8 * B3
B ~ 0.5 * A
"
get_model_names(m)
}
