% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{lav2ram}
\alias{lav2ram}
\title{Extract standardized RAM matrices from lavaan object}
\usage{
lav2ram(fit)
}
\arguments{
\item{fit}{An object of class lavaan}
}
\value{
list of RAM matrices A (assymetric paths), S (symmetric paths), and F (filter matrix)
}
\description{
Extract standardized RAM matrices from lavaan object
}
